% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfascoring-package.R
\docType{package}
\name{cfascoring-package}
\alias{cfascoring}
\alias{cfascoring-package}
\title{cfascoring: CFA Scoring Service}
\description{
Tools to score CFA models and interact with Azure.
}
\keyword{internal}
