% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{download_file_from_container}
\alias{download_file_from_container}
\title{Download specified blobs from Blob Storage and save them in a local dir}
\usage{
download_file_from_container(
  blob_storage_path,
  local_file_path,
  storage_container
)
}
\arguments{
\item{blob_storage_path}{A character of a blob in \code{storage_container}}

\item{local_file_path}{The local path to save the blob}

\item{storage_container}{The blob storage container with \code{blob_storage_path}}
}
\value{
Invisibly, \code{local_file_path}
}
\description{
Download specified blobs from Blob Storage and save them in a local dir
}
\seealso{
Other azure:
\code{\link{download_if_specified}()},
\code{\link{fetch_blob_container}()},
\code{\link{upload_to_container}()}
}
\concept{azure}
