% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{upload_to_container}
\alias{upload_to_container}
\title{Upload to container}
\usage{
upload_to_container(blob_path, blob_storage_container)
}
\arguments{
\item{blob_path}{The local path of the blob to upload}

\item{blob_storage_container}{The name of the container to upload to}
}
\value{
The path of the file
}
\description{
Upload to container
}
\seealso{
Other azure:
\code{\link{download_file_from_container}()},
\code{\link{download_if_specified}()},
\code{\link{fetch_blob_container}()}
}
\concept{azure}
