% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{write_scores}
\alias{write_scores}
\title{Write scores}
\usage{
write_scores(predicted, report_date, final, output)
}
\arguments{
\item{predicted}{modeled predictions in Epinow2 pipeline format}

\item{report_date}{The vintage of the data to which the model was fit.
Currently only used to check that the vintage is not too
recent relative to the final data.}

\item{final}{Finalized, observed data. Usually the latest available vintage.}

\item{output}{Path to write scored estimates in parquet format}
}
\value{
The output path of the file
}
\description{
Write scores
}
