% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{fetch_blob_container}
\alias{fetch_blob_container}
\title{Load Azure Blob container using managed identity credentials}
\usage{
fetch_blob_container(container_name)
}
\arguments{
\item{container_name}{The Azure Blob Storage container associated with the
credentials}
}
\value{
A Blob endpoint
}
\description{
As a result it is an impure function, and should be used bearing that
warning in mind.
}
\seealso{
Other azure:
\code{\link{download_file_from_container}()},
\code{\link{download_if_specified}()}
}
\concept{azure}
